requirements.txt
Pillow==5.2.0
PySDL2==0.9.7
